backend/requirements.txt
fastapi
openai
uvicorn
black
tqdm
anthropic
firebase-admin
python-dotenv
websockets
pydantic
pyyaml
stripe
pytest
pytest
pytest-asyncio
pytest-mock
python-multipart
instructor
pypubsub
motor
beanie